requirements.txt
fastapi
uvicorn
scikit-learn
numpy
pydantic
python-multipart
aiohttp
python-dotenv
